+++ comb_spec_searcher/rule_db/base.py
@@ -3,35 +3,25 @@ A database for rules.
 """
 import abc
 from collections import defaultdict
-from typing import (
-    Dict,
-    Iterable,
-    Iterator,
-    List,
-    MutableMapping,
-    Optional,
-    Set,
-    Tuple,
-    Union,
-    cast,
-)
+from typing import Dict, Iterable, Iterator, List, MutableMapping, Optional, Set, Tuple
 
 from logzero import logger
 
-from .class_db import ClassDB
-from .equiv_db import EquivalenceDB
-from .exception import SpecificationNotFound, StrategyDoesNotApply
-from .strategies import Rule, StrategyPack, VerificationRule
-from .strategies.rule import AbstractRule
-from .strategies.strategy import AbstractStrategy, StrategyFactory
-from .tree_searcher import (
+from comb_spec_searcher.equiv_db import EquivalenceDB
+from comb_spec_searcher.exception import SpecificationNotFound
+from comb_spec_searcher.strategies import AbstractStrategy, Rule, VerificationRule
+from comb_spec_searcher.strategies.rule import AbstractRule
+from comb_spec_searcher.tree_searcher import (
     Node,
     iterative_proof_tree_finder,
     iterative_prune,
     proof_tree_generator_dfs,
     prune,
     random_proof_tree,
+    smallish_random_proof_tree,
 )
+
+__all__ = ["RuleDBBase", "RuleDB"]
 
 Specification = Tuple[List[Tuple[int, AbstractStrategy]], List[List[int]]]
 RuleKey = Tuple[int, Tuple[int, ...]]
@@ -149,7 +139,9 @@ class RuleDBBase(abc.ABC):
                 return start, ends
         return None
 
-    def find_specification(self, label: int, iterative: bool = False) -> Node:
+    def find_specification(
+        self, label: int, minimization_time_limit: float, iterative: bool = False
+    ) -> Node:
         """Search for a specification based on current data found."""
         rules_dict = self.rules_up_to_equivalence()
         # Prune all unverified labels (recursively)
@@ -169,20 +161,26 @@ class RuleDBBase(abc.ABC):
                     rules_dict, root=self.equivdb[label]
                 )
             else:
-                specification = random_proof_tree(rules_dict, root=self.equivdb[label])
+                specification = smallish_random_proof_tree(
+                    rules_dict, self.equivdb[label], minimization_time_limit
+                )
         else:
             raise SpecificationNotFound("No specification for label {}".format(label))
         return specification
 
     def get_specification_rules(
-        self, label: int, iterative: bool = False
+        self, label: int, minimization_time_limit: float, iterative: bool = False
     ) -> Specification:
         """
         Return a list of pairs (label, rule) which form a specification.
         The specification returned is random, so two calls to the function
         may result in a a different output.
         """
-        proof_tree_node = self.find_specification(label=label, iterative=iterative)
+        proof_tree_node = self.find_specification(
+            label=label,
+            iterative=iterative,
+            minimization_time_limit=minimization_time_limit,
+        )
         return self._get_specification_rules(label, proof_tree_node)
 
     def _get_specification_rules(
@@ -287,83 +285,3 @@ class RuleDB(RuleDBBase):
     @property
     def rule_to_strategy(self) -> Dict[Tuple[int, Tuple[int, ...]], AbstractStrategy]:
         return self._rule_to_strategy
-
-
-class RecomputingDict(MutableMapping[RuleKey, AbstractStrategy]):
-    """
-    A mapping from rules to strategies that recompute the strategy every time it's
-    needed instead of storing it in order to save memory.
-
-    Also in order to save memory we store flat version of the rules, i.e. for a rule
-    (a, (b, c,...)) we store (a, b, c,...)
-    """
-
-    def __init__(self, classdb: ClassDB, strat_pack: StrategyPack) -> None:
-        self.classdb = classdb
-        self.pack = strat_pack
-        self.rules: Set[Tuple[int, ...]] = set()
-
-    @staticmethod
-    def _flatten(tuple_: RuleKey) -> Tuple[int, ...]:
-        return (tuple_[0],) + tuple_[1]
-
-    @staticmethod
-    def _unflatten(tuple_: Tuple[int, ...]) -> RuleKey:
-        return (tuple_[0], tuple_[1:])
-
-    def __getitem__(self, key: RuleKey) -> AbstractStrategy:
-        if self._flatten(key) not in self.rules:
-            raise KeyError(key)
-        parent = self.classdb.get_class(key[0])
-        for strat in self.pack:
-            if isinstance(strat, StrategyFactory):
-                strats_or_rules: Iterable[Union[Rule, AbstractStrategy]] = strat(parent)
-            else:
-                strats_or_rules = [strat]
-            rules: Iterator[AbstractRule] = (
-                x(parent) if isinstance(x, AbstractStrategy) else x
-                for x in strats_or_rules
-            )
-            for rule in rules:
-                try:
-                    nonempty_children = tuple(
-                        c for c in rule.children if not self.classdb.is_empty(c)
-                    )
-                    end_labels = tuple(
-                        sorted(map(self.classdb.get_label, nonempty_children))
-                    )
-                    if end_labels == key[1]:
-                        return rule.strategy
-                except StrategyDoesNotApply:
-                    pass
-        err_message = (
-            f"Could not recompute the strategy for the rule {key} with "
-            " any of the strategies"
-        )
-        raise RuntimeError(err_message)
-
-    def __setitem__(self, key: RuleKey, value: AbstractStrategy) -> None:
-        self.rules.add(self._flatten(key))
-
-    def __delitem__(self, key: RuleKey) -> None:
-        self.rules.remove(self._flatten(key))
-
-    def __iter__(self) -> Iterator[RuleKey]:
-        for rule in self.rules:
-            yield self._unflatten(rule)
-
-    def __len__(self) -> int:
-        return len(self.rules)
-
-    def __contains__(self, key: object) -> bool:
-        return self._flatten(cast(RuleKey, key)) in self.rules
-
-
-class RuleDBForgetStrategy(RuleDBBase):
-    def __init__(self, classdb: ClassDB, strat_pack: StrategyPack) -> None:
-        super().__init__()
-        self.rules = RecomputingDict(classdb, strat_pack)
-
-    @property
-    def rule_to_strategy(self):
-        return self.rules